requirements.txt
accelerate==1.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.1.0
avidtools==0.1.2
babel==2.17.0
backoff==2.2.1
base2048==0.1.3
beautifulsoup4==4.13.3
bert-score==0.3.13
bitsandbytes==0.45.2
bleach==6.2.0
boto3==1.37.6
botocore==1.37.6
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cmd2==2.4.3
cohere==4.57
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.16.1
DateTime==5.5
debugpy==1.8.13
decorator==5.2.1
deepl==1.17.0
deepspeed==0.16.3
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
dnspython==2.7.0
ecoji==0.1.1
einops==0.8.1
emoji==2.14.1
eval_type_backport==0.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastavro==1.10.0
fastchat==0.1.0
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fschat==0.2.36
fsspec==2023.10.0
# Editable install with no version control (garak==0.10.3.post1)
-e /root/autodl-tmp/lxy/gitcode/garak
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.21.1
googleapis-common-protos==1.69.0
greenlet==3.1.1
h11==0.14.0
hjson==3.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
hydra-core==1.3.2
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.11
latex2mathml==3.77.0
litellm==1.62.1
lorem==0.1.1
lxml==5.3.2
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nemollm==0.3.5
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
ninja==1.11.1.3
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
nvdlib==0.7.9
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
octoai-sdk==0.10.1
olefile==0.47
ollama==0.4.7
omegaconf==2.3.0
openai==1.65.3
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
ply==3.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydevd==3.3.0
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.4.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-futures==1.0.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
s3transfer==0.11.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.0.2
sentencepiece==0.2.0
setuptools==75.8.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
stdlibs==2024.12.3
svgwrite==1.4.3
sympy==1.13.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20250301
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.17.2
unstructured-client==0.32.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.13
wn==0.9.5
wrapt==1.17.2
xdg-base-dirs==6.0.2
xxhash==3.5.0
yarl==1.18.3
zalgolib==0.2.2
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
